Added Suggested CPs and Email Filtering
Added a new class, EmailParser, to filter through and find the S Request email in Outlook for the user.
In EmailParser, I added a function that finds the CPs inside the S Request email to pass onto the other classes.

In ExcelReader, I modified code I added a week or so ago to be more functionally appropriate. I filtered more junk from the cp_names list on first pass-through, so there are no more blank " " strings in the list.

Nothing changed in ExcelPlanner, I just saved the file when I tabbed into it by habit.

In main, I added a new function that compares the new suggested_cps from EmailParser with the cps found from ExcelReader.
I also reorganized how the questions are asked, so the same amount of questions are asked, just in a different order to maintain simplicity throughout the file.