Fixed index out of bounds issue
Fixed an issue where the CFA had an extra space in it on every line, so I made it handle all possible spaces.

I added some functionality in my formatting list where it will always be the size of the first row in the excel_data list, which the first list should always be the length of every list. If it is not, it will crash again. I will get to that when it happens, but this is a very unlikely situation.