Added functionality to format the codeline name
In ExcelReader, I added some functionality in the function that handles grabbing the codeline information.

In regards to our testing on the UP project, the excess of information in that specific field is not necessary for us. This being the case, I have added a couple lines to shorten that type name.