Bug fix for finding CP name
On line 252 of ExcelReader.py, I added "and position_x > 0" to the line.

This helps prevent any errors when finding a different occurrence of the CP name that does not help find relevant information. The CP name must never be used from the first column in a row. It must always be used from the second column. Since the program will take the first occurrence of the CP name and run with that, I have not added error handling for finding another occurrence after the second, as it is not necessarily required.